Refactor tests for datasets to be in their own folder
